refactor(trusty-search): split service/reindex into focused submodules (closes #1175)
Split the 2975-line service/reindex/mod.rs monolith into 9 focused submodules,
each under the 500-SLOC production cap:

- progress.rs     — ReindexProgress, ReindexStatus, broadcast/replay constants
- semaphore.rs    — interactive + background semaphores, queue-depth counter
- hash.rs         — per-batch SHA-256 content hashing
- stages.rs       — stage-transition helpers (lexical→semantic→graph pipeline)
- batch.rs        — single-batch parse/embed/commit cycle
- completion.rs   — KG rebuild + terminal 'complete' SSE event builder
- corpus_swap.rs  — atomic index.redb.tmp → index.redb swap
- guard.rs        — ReindexTerminationGuard RAII safety guard
- orchestrator.rs — top-level spawn_reindex / spawn_reindex_with_cleanup

mod.rs is reduced to a thin re-export facade so all external
crate::service::reindex::* paths resolve unchanged.

Also updates .line-cap-allowlist.tsv to remove the old 1658-SLOC
grandfathered budget (now 35 SLOC).
